Create NavBar for header, and ContextProvider to avoid the problem of prop drilling
In this commit a navbar was created with the name of the application and
a log in button. The next step we took was creating a context provider that
will make a global state and make our state available to the children
components, and also makes our reducer function available to the entire
app aka chilren of context provider. This was done to that we can move
away from prop drilling since that clutters up our code. So after this
all of the children components now have a way to access state and can
change it by calling the useContext function.