Create drop down menu once profile picture is clicked. implement logout feature
Once the user clicks the picture a drowdown will show up that will allow the
user to view there profile or logout. Logout has been implemented calling teh same
action creator and reducer as login;however, we will be passing a null value instead.
In this change we have alo implemented another check for the ternary operation
to show up the icons, this is because when we update state and then remove it, the nested
state remains with the null value which means we have to use user.user == null to continue
to check if the user is logged in or not.