Create actions and reducer for state of login form. Move usermenu into users folder
actions and reducer was created for login form so we can manage the state of
the form globally without much prop drilling. Two function were created for the
login form, one to open the form which sets the value to true, and one is to
close the form which will set the value to false.