refactor login state in redux. Create login component which displays form that user will use to login, add a name field in it currently
login name was confusing earlier but I now refactored it to contain a
better named action and reducer. Login component was made and connected
to the redux store to allow us to dispatch the action creators for login.
Form field was also created to allow the user to only currently add a name
in the form.