Create login form. refactor reducer and action to mimic(reminder: state in index is just state in reducer). Create passwordfield componenet since it will be used twice
Problem with redux was encountered that was resulting us to store values
into the state object in both user and login. Reminder that the state in
the reducer is the name for the name of the reducer in the index file.
We can also nest elements into object as was done previously to expand On
the reducer.