Switch between login and register form. Change title of form using useeffect
Added a state variable that once altered will give us some extra input
fields in teh register form. The register form also contains a different
name from the login form this is happening due to us altering the boolean
attribute of isregister and the useeffect hook running due to this change.
This useEffect will alter the naem of our title use the sethooks and
resutsl in a change between forms. The next step is to introduce a google
login button to sign the user in.