Get project client ID and import it into client environment. Use google window object to prompt google login. Reenable google log in button once exited or timed out. Import JWT decoder to decode json response from the google object
The handleGoogleLogin function is responsible for initializing the Google
Sign In API with the provided client ID, and also triggers the prompt for
the user to sign in. If the user is not displayed a notification to sign in,
the function sets an expired cookie to clear the current state, and prompts
the user again. If the user is skipped or dismissed the sign-in prompt, the function sets the disabled property to false.
The handleResponse function is the callback function that is triggered
once the user successfully signs in. This function decodes the JWT token
received in the response, extracts the user information such as email, name,
profile picture, and token. Finally, it updates the user information in the state
and closes the login dialog. If there is an error in the try block, the catch block is
executed and logs the error, and also dispatches an action to show an error alert with
the error message.

UserIcon component was also alterd to fix the property of the user object to its correct name
from photoUrl -> photoURL. Now because of this the image of the users gmail will display as
an icon