Create creat api user call in backend that takes user registered data from form. Ensure the user doesn't exists in database using email and store it in user collection.
Created a createUser action creator that will call the creatUser api function that will
send a request to the backend along with the data we want to send about the newly
created user. This data would be used in the creatUser controller to figure out if the
user exists and if they don't they will be stored in the database. When we call the api
creat method we dispatch our loader action that starts the loader up and we close it once we
receive the request back with the newly created user data which we then store into the
current user state using the updateUser dispatcher and save it into the local storage. We
also close the login if this is successful, if successful we dispatch success alert in
the state. If not successful we dispatch not success alert using the error message we
received from the backend