Creat get api for the retrieval of user when logging in from the front end. Creat getuser action creator and api call.
getuser api function will retrieve the the corresponding user based on there email addres. The function will also
compare the inputted password encrypt it and compare it to the one stored in our database. The function will then return
all the user information where we then dispatch the updateuser function which will update that user state.