Create profile update action in both frontend and back that allows user to change name and profile picture.
User that are not signed it with google can now alter there profile picture
and there name. Connected the updatepicture action with firebase for image to be stored and custom
image urls to be created based on the uploaded picture and the name of the user. The name along with the
iamge url are sent to the backend to be assigned and overwrite the previous users data. Then is returend and overwrites the
current users information and closes the profile prompt