Create add room folder with 3 components called addDetails, addLocation, and addRoom
These components will be created so the user who is trying to upload a room to the website
will be able to to do step by step using these 3 component. The displayed instructions
will depend on the activesteps state which also determines which step of the stepper
the user is on. This was implemented when the box was added to our
AddRoom Component.