Create CircularProgressWithLabel, ProgressItem, and ProgressList components. Add new state element called images to store images dropped by user into state.
CircularProgressWithLabel: This component is a circular progress bar that shows the progress of a file upload. It displays the percentage of progress in the center of the progress bar.

ProgressItem: This component represents a single item in a list of image uploads. It shows an image preview of the file being uploaded, and a progress bar that displays
the progress of the upload. It also dispatches an action to update the images state in the Redux store when the upload is complete.

ProgressList: This component is a list of ProgressItem components. It takes an array of file objects as a prop and renders a ProgressItem component for each file in the array.
The ProgressList component is responsible for rendering the list in a grid format.

image state was added and reducer to add image was created so dropped user images are held until uploaded.