Implement the ability to delete images placed by the user. Create ImagesList component and deleteFile method for firebase. Implement new reducer case in images.
The ImagesList component receives the currentUser and images props from the Redux store using the mapStateToProps function.
The currentUser prop is an object that contains information about the current user, while the images prop is an array of URLs that point to the images to be displayed.
This component also uses the useDispatch hook from the react-redux library to dispatch actions to the Redux store. In particular, the handleDelete function is called when the delete
button is clicked for an image. This function dispatches a DELETE_IMAGE action to the Redux store with the image URL as the payload. It then uses the deleteFile function to delete
the image file from Firebase Storage.

The deleteFile function exports a deleteFile function that takes a filePath argument and deletes the corresponding file from Firebase Storage. It uses the ref and deleteObject
functions from the Firebase Storage SDK to locate and delete the file.

Implement the ImagesList component inside of the AddImages component so it is shown on the UI. DELETE_IMAGE reducer created and called when x button is clicked on image.