Create title and description step for rooms in room card creation that has instance check and typing icon.
This was done by implementing a component called InfoField that would be inserted into the AddDetails component to add
the skeleton of a text field that we then used with the title as well as the room description. Everytime a user types
a symbol will pop up that shows the user typing, when they stop it will notify the user if the text is too short, or too long.