Implement mapbox in step 1 of room setup. Import mapbox token and add location state in reducers
This will allow the user to designate where he/she wants the room to be.
The map will be imported using mapbox access token that wil show a geographical
map with a marker that will move and at the same time alter teh longitude, and
latitude in the location state in redux.