Allow mapbox to locate the city of the user by using mapboxes geolocator
When the user refreshes the page the marker will start out at there city.
This is because we are using the geolocator component that is passed the
prop trackuserlocation. Once this is called the lng and lat var in the
current compnent are updates to the corresponding city and this is then
dispatched to the redux state which updates the lng and lat in the global
store. Since our marker uses these values for its position on the map it
will move to the corresponding city