Handle the state of the submit button showing when all three steps of add room are completed
This was done so the user cannot prematurely add a room into the server and display it before
they finish all details of the room. If one of the steps isn't complete the submit button will not
appear and the room may not be added. However, if they have all three complete they may add the room
*Submit functionality is in progress*