Implement CreateRoom action that will call the creatRoom API in backend.
The CreatRoom function will receive data from the AddRoom component after
the submit button is clicked and it will recieve the room object, along with
the currentUser to check for the token, and the dispatch to dispatch
reducers. The CreateRoom action will use a middleware fetchData function
that will handle the call and return it back to the CreatRoom function. Once
it gets a response back the room will be stored in the database and an
alert will be made to ensure the room was added.

The createRoom function receives three arguments: room, currentUser,
and dispatch, and it dispatches actions to a Redux store to handle loading
and alert messages. It also calls the fetchData function, passing in the
url, room, currentUser?.token (if currentUser is defined),
and dispatch arguments, to make an HTTP request to the server.