Reset state of room information after user submits room
This was done so when the user has completed the creation of there room
they are notified of the results and taken back to the home screen. Some modifications
to the details, images, and locations reducers were also implemented with the
identical action type cases so we can reset these back to normal when the is
done creating there room. This is because when they go to make another room
there previous information isn't already filled in.