Prevent unathenticated and anonymouse users from access add room page by adding a forbidden access Warning
This was done by creating a protected component around the add room
page that will identify if the currentUser state is not empty in the redux
state. If it is emtpy then the protected component will not load its children component
which in this case it addRoom. Instead it will load the AccessMessage component which is a warning
banner that forbids access and allows users to login or create an account first.