Create get room controller and route
The getRoom controller will be the endpoint in our server which will return
the list of rooms in our room collection database. This will send a response
of succss and rooms as the payload. Using express we also create a get router
that leads to the getRooms controller when accessed.