Create action creator for getRoom, and upate room reducers
Created getRoom action creator function that will send a request to our
server and hit the getRoom function that will retrieve all the rooms in
our database and send them to the front end. The getRoom action creator
will receive these rooms and populat the rooms state array in redux
with all the received rooms.