Create dashboard page
This was created by clearin the app.js file of all components and moving them
to the Home file in pages. Then router was used to direct user to proper
page based on route. The creation of a dashboard component was also made
with a MUI template where we will build our dashboard.