Implement light and dark mode feature in user dashboard
This was implemented by wrapping our component in a theme that would
alter based on a created state called dark if true it would be dark, and
if false it would be light. This was added so users who prefer less light
in there eye would have a better time on the application.